refined app/customers/add page.tsx
moved customerSchema to app/schemas/index.ts
moved checkAndCreatePolicyHolder to app/actions/checkAndCreatePolicyHolder.tsx
