+++ app/schemas/index.ts
@@ -20,3 +20,11 @@ export const RegisterSchema = z.object({
     message: "Name is required",
   }),
 });
+
+export const customerSchema = z.object({
+  id: z.string().nonempty("NRIC is required"),
+  email: z.string().email("Invalid email").nonempty("Email is required"),
+  firstName: z.string().nonempty("First Name is required"),
+  lastName: z.string().nonempty("Last Name is required"),
+  policy: z.string().nonempty("Please select a policy"),
+});
